+++ Source/Scene/CentralBodySurface.js
@@ -796,6 +796,8 @@ define([
     var northeastScratch = new Cartesian3();
 
     function createRenderCommandsForSelectedTiles(surface, context, frameState, shaderSet, projection, centralBodyUniformMap, commandList, renderState) {
+        displayCredits(surface, frameState);
+
         var viewMatrix = frameState.camera.viewMatrix;
 
         var maxTextures = context.getMaximumTextureImageUnits();
@@ -974,6 +976,14 @@ define([
                         }
                         applyGamma = applyGamma || uniformMap.dayTextureOneOverGamma[numberOfDayTextures] !== 1.0 / ImageryLayer.DEFAULT_GAMMA;
 
+                        if (defined(imagery.credits)) {
+                            var creditDisplay = frameState.creditDisplay;
+                            var credits = imagery.credits;
+                            for (var creditIndex = 0, creditLength = credits.length; creditIndex < creditLength; ++creditIndex) {
+                                creditDisplay.addCredit(credits[creditIndex]);
+                            }
+                        }
+
                         ++numberOfDayTextures;
                     }
 
@@ -1044,6 +1054,31 @@ define([
             }
             tile.meshForWireframePromise = undefined;
         });
+    }
+
+    function displayCredits(surface, frameState) {
+        var creditDisplay = frameState.creditDisplay;
+        var credit;
+
+        if (surface._terrainProvider.isReady()) {
+            credit = surface._terrainProvider.getCredit();
+            if (defined(credit)) {
+                creditDisplay.addCredit(credit);
+            }
+        }
+
+        var imageryLayerCollection = surface._imageryLayerCollection;
+        for ( var i = 0, len = imageryLayerCollection.getLength(); i < len; ++i) {
+            var layer = imageryLayerCollection.get(i);
+            if (layer.show) {
+                if (layer.getImageryProvider().isReady()) {
+                    credit = layer.getImageryProvider().getCredit();
+                    if (defined(credit)) {
+                        creditDisplay.addCredit(credit);
+                    }
+                }
+            }
+        }
     }
 
     return CentralBodySurface;